requirements.txt
torch
torchvision
streamlit
opencv-python
Pillow
numpy
requests
